インターフェースクラスからシングルトンへ改修

+++ glfw3_app/common/img_io/bmp_io.cpp
@@ -169,8 +169,8 @@ namespace img {
 			bmp.depth  = mgetwl(bih + BIH_WBITCOUNT);
 			bmp.compression   = mgetdwl(bih + BIH_DCOMPRESSION);
 			bmp.palette_size  = RGBQUAD_SIZE;
-			if(bmp.height < 0) {
-				bmp.height  = -bmp.height;
+			if(static_cast<int>(bmp.height) < 0) {
+				bmp.height  = -static_cast<int>(bmp.height);
 				bmp.topdown = true;
 			}
 		} else {
@@ -485,9 +485,9 @@ namespace img {
 		for( ; ; ) {
 			unsigned int reclen;
 			while(bfcnt < (reclen = 2) || (bfptr[0] == 0 && (
-					(bfptr[1] == 2 && bfcnt < (reclen += 2)) ||
-					 bfptr[1] >= 3 && bfcnt < (reclen += (bfptr[1] * bmp.depth + 15) / 16 * 2)
-					))) {
+					 (bfptr[1] == 2 && bfcnt < (reclen += 2)) || bfptr[1] >= 3 &&
+					 bfcnt < (reclen += (bfptr[1] * bmp.depth + 15) / 16 * 2)
+												   ))) {
 				if(bfptr != buf && bfcnt != 0) memmove(buf, bfptr, bfcnt);
 				size_t rd = fin.read(buf + bfcnt, 1, sizeof(buf) - bfcnt);
 				if(rd == 0) {
